+++ package.json
@@ -12,7 +12,7 @@
     "test": "vitest"
   },
   "dependencies": {
-    "@420cry/420cry-lib": "=0.2.7",
+    "@420cry/420cry-lib": "=0.3.0",
     "@heroicons/react": "^2.2.0",
     "autoprefixer": ">= 10.4.21",
     "axios": ">= 1.12.0",
