+++ src/components/common/header/ThemeToggle.tsx
@@ -2,6 +2,7 @@
 
 import { JSX } from 'react'
 import { useTheme } from '@/lib'
+import { SunIcon, MoonIcon } from '@420cry/420cry-lib'
 
 const ThemeToggle = (): JSX.Element => {
   const { theme, toggleTheme } = useTheme()
@@ -9,19 +10,11 @@ const ThemeToggle = (): JSX.Element => {
   return (
     <div className="flex items-center justify-center gap-2">
       {/* Sun Icon */}
-      <svg
+      <SunIcon
         className={`w-4 h-4 transition-colors duration-300 ease-in-out ${
           theme === 'light' ? 'text-yellow-500' : 'text-gray-400'
         }`}
-        fill="currentColor"
-        viewBox="0 0 20 20"
-      >
-        <path
-          fillRule="evenodd"
-          d="M10 2a1 1 0 011 1v1a1 1 0 11-2 0V3a1 1 0 011-1zm4 8a4 4 0 11-8 0 4 4 0 018 0zm-.464 4.95l.707.707a1 1 0 001.414-1.414l-.707-.707a1 1 0 00-1.414 1.414zm2.12-10.607a1 1 0 010 1.414l-.706.707a1 1 0 11-1.414-1.414l.707-.707a1 1 0 011.414 0zM17 11a1 1 0 100-2h-1a1 1 0 100 2h1zm-7 4a1 1 0 011 1v1a1 1 0 11-2 0v-1a1 1 0 011-1zM5.05 6.464A1 1 0 106.465 5.05l-.708-.707a1 1 0 00-1.414 1.414l.707.707zm1.414 8.486l-.707.707a1 1 0 01-1.414-1.414l.707-.707a1 1 0 011.414 1.414zM4 11a1 1 0 100-2H3a1 1 0 000 2h1z"
-          clipRule="evenodd"
-        />
-      </svg>
+      />
 
       {/* Toggle Button */}
       <button
@@ -37,15 +30,11 @@ const ThemeToggle = (): JSX.Element => {
       </button>
 
       {/* Moon Icon */}
-      <svg
+      <MoonIcon
         className={`w-4 h-4 transition-colors duration-300 ease-in-out ${
           theme === 'dark' ? 'text-blue-400' : 'text-gray-400'
         }`}
-        fill="currentColor"
-        viewBox="0 0 20 20"
-      >
-        <path d="M17.293 13.293A8 8 0 016.707 2.707a8.001 8.001 0 1010.586 10.586z" />
-      </svg>
+      />
     </div>
   )
 }
