+++ src/components/modal/transactions/XPUBTransactionModal.tsx
@@ -17,7 +17,7 @@ import {
   TooltipItem,
 } from 'chart.js'
 import 'chartjs-adapter-date-fns'
-import { CryModal } from '@420cry/420cry-lib'
+import { CryModal, ChartLineIcon, CopyIcon } from '@420cry/420cry-lib'
 
 ChartJS.register(
   LineElement,
@@ -185,24 +185,12 @@ export default function XPUBTransactionModal({
       onClose={onClose}
       title={`${t('title')} (XPUB)`}
       ariaLabelledBy="xpub-transaction-modal-title"
-      maxWidth="max-w-3xl"
+      maxWidth="3xl"
     >
       <section className="mb-8">
         <div className="flex items-center gap-3 mb-6">
           <div className="w-8 h-8 bg-blue-100 dark:bg-blue-900/30 rounded-lg flex items-center justify-center">
-            <svg
-              className="w-4 h-4 text-blue-600 dark:text-blue-400"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"
-              />
-            </svg>
+            <ChartLineIcon className="w-4 h-4 text-blue-600 dark:text-blue-400" />
           </div>
           <h3 className="text-xl font-bold text-gray-900">
             {t('balanceGraph.title', { defaultValue: 'Balance Over Time' })}
@@ -218,19 +206,7 @@ export default function XPUBTransactionModal({
       <section>
         <div className="flex items-center gap-3 mb-4">
           <div className="w-8 h-8 bg-purple-100 dark:bg-purple-900/30 rounded-lg flex items-center justify-center">
-            <svg
-              className="w-4 h-4 text-purple-600 dark:text-purple-400"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M9 5H7a2 2 0 00-2 2v10a2 2 0 002 2h8a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2m-3 7h3m-3 4h3m-6-4h.01M9 16h.01"
-              />
-            </svg>
+            <CopyIcon className="w-4 h-4 text-purple-600 dark:text-purple-400" />
           </div>
           <h3 className="text-xl font-bold text-gray-900">{t('title')}</h3>
         </div>
