+++ src/components/pages/2fa/alternative/TwoFactorAlternativeSendForm.tsx
@@ -1,7 +1,13 @@
 'use client'
 
 import { JSX, useState, useMemo, useEffect } from 'react'
-import { CryButton, CryTextField } from '@420cry/420cry-lib'
+import {
+  CryButton,
+  CryTextField,
+  CheckCircleIcon,
+  LoadingIcon,
+  RefreshIcon,
+} from '@420cry/420cry-lib'
 import { useTranslations } from 'next-intl'
 import {
   HOME_ROUTE,
@@ -178,19 +184,7 @@ const TwoFactorAlternativeSendForm = (): JSX.Element => {
     <div className="w-full max-w-lg bg-gradient-to-br from-white to-gray-50 shadow-2xl rounded-3xl p-12 border border-gray-200/50 backdrop-blur-sm">
       <div className="text-center mb-8">
         <div className="w-16 h-16 bg-gradient-to-br from-green-500 to-emerald-600 rounded-2xl mx-auto mb-6 flex items-center justify-center shadow-lg">
-          <svg
-            className="w-8 h-8 text-white"
-            fill="none"
-            stroke="currentColor"
-            viewBox="0 0 24 24"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
-            />
-          </svg>
+          <CheckCircleIcon className="w-8 h-8 text-white" />
         </div>
         <h1 className="text-3xl font-bold mb-4 bg-gradient-to-r from-gray-900 to-gray-700 bg-clip-text text-transparent">
           {title}
@@ -259,19 +253,7 @@ const TwoFactorAlternativeSendForm = (): JSX.Element => {
             shape="rounded"
           >
             <div className="flex items-center justify-center space-x-2">
-              <svg
-                className="w-5 h-5"
-                fill="none"
-                stroke="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
-                />
-              </svg>
+              <CheckCircleIcon className="w-5 h-5" />
               <span>{t('2fa.verify.verifyCode')}</span>
             </div>
           </CryButton>
@@ -289,24 +271,12 @@ const TwoFactorAlternativeSendForm = (): JSX.Element => {
             <div className="flex items-center justify-center space-x-2">
               {canResend ? (
                 <>
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-                    />
-                  </svg>
+                  <RefreshIcon className="w-5 h-5 rotate-180" />
                   <span>{t('2fa.alternative.resendCode')}</span>
                 </>
               ) : (
                 <>
-                  <div className="w-5 h-5 border-2 border-white border-t-transparent rounded-full animate-spin"></div>
+                  <LoadingIcon className="w-5 h-5 animate-spin" />
                   <span>{`${t('2fa.alternative.resendIn')} ${minutes}:${seconds
                     .toString()
                     .padStart(2, '0')}`}</span>
