+++ src/lib/context/notification/NotificationRenderer.tsx
@@ -3,6 +3,13 @@
 import { JSX, useEffect, useState } from 'react'
 import { createPortal } from 'react-dom'
 import { useNotification } from './NotificationContext'
+import {
+  SuccessCircleIcon,
+  WarningTriangleIcon,
+  InfoIcon,
+  XIcon,
+  ErrorCircleIcon,
+} from '@420cry/420cry-lib'
 
 export function NotificationRenderer(): JSX.Element | null {
   const { notifications, removeNotification } = useNotification()
@@ -17,69 +24,13 @@ export function NotificationRenderer(): JSX.Element | null {
   const getIcon = (type: string) => {
     switch (type) {
       case 'success':
-        return (
-          <svg
-            className="w-6 h-6"
-            fill="none"
-            stroke="currentColor"
-            viewBox="0 0 24 24"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
-            />
-          </svg>
-        )
+        return <SuccessCircleIcon className="w-6 h-6" />
       case 'error':
-        return (
-          <svg
-            className="w-6 h-6"
-            fill="none"
-            stroke="currentColor"
-            viewBox="0 0 24 24"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z"
-            />
-          </svg>
-        )
+        return <ErrorCircleIcon className="w-6 h-6" />
       case 'warning':
-        return (
-          <svg
-            className="w-6 h-6"
-            fill="none"
-            stroke="currentColor"
-            viewBox="0 0 24 24"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-            />
-          </svg>
-        )
+        return <WarningTriangleIcon className="w-6 h-6" />
       case 'info':
-        return (
-          <svg
-            className="w-6 h-6"
-            fill="none"
-            stroke="currentColor"
-            viewBox="0 0 24 24"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-            />
-          </svg>
-        )
+        return <InfoIcon className="w-6 h-6" />
       default:
         return null
     }
@@ -173,19 +124,7 @@ export function NotificationRenderer(): JSX.Element | null {
                   className={`flex-shrink-0 ${colors.icon} ${colors.closeHover} rounded-lg p-1 transition-colors duration-200`}
                   aria-label="Close notification"
                 >
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M6 18L18 6M6 6l12 12"
-                    />
-                  </svg>
+                  <XIcon className="w-5 h-5" />
                 </button>
               </div>
             </div>
