+++ src/lib/hooks/useCurrencyPreference.ts
@@ -3,7 +3,14 @@
 import { useState, useEffect } from 'react'
 import { CurrencyService, CurrencyCode } from '@/lib'
 
-export function useCurrencyPreference() {
+export function useCurrencyPreference(): {
+  selectedCurrency: CurrencyCode
+  changeCurrency: (currency: CurrencyCode) => void
+  formatAmount: (amount: number) => string
+  getSymbol: () => string
+  isLoading: boolean
+  availableCurrencies: CurrencyCode[]
+} {
   const [selectedCurrency, setSelectedCurrency] = useState<CurrencyCode>('USD')
   const [isLoading, setIsLoading] = useState(true)
 
@@ -14,16 +21,16 @@ export function useCurrencyPreference() {
     setIsLoading(false)
   }, [])
 
-  const changeCurrency = (currency: CurrencyCode) => {
+  const changeCurrency = (currency: CurrencyCode): void => {
     setSelectedCurrency(currency)
     CurrencyService.setPreferredCurrency(currency)
   }
 
-  const formatAmount = (amount: number) => {
+  const formatAmount = (amount: number): string => {
     return CurrencyService.formatCurrency(amount, selectedCurrency)
   }
 
-  const getSymbol = () => {
+  const getSymbol = (): string => {
     return CurrencyService.getCurrencySymbol(selectedCurrency)
   }
